Adding a custom core ml
- Added python script to cliassify images in folders
- Added core ml model generated by classifier
- Added requirements.txt to store the requirements for python script
- Added readme file with description of the project and instructions
