--- UnoBeer/Assets.xcassets/wmb.imageset/Contents.json
@@ -1,23 +1,0 @@
-{
-  "images" : [
-    {
-      "idiom" : "universal",
-      "filename" : "wmb125.png",
-      "scale" : "1x"
-    },
-    {
-      "idiom" : "universal",
-      "filename" : "Untitled-1.png",
-      "scale" : "2x"
-    },
-    {
-      "idiom" : "universal",
-      "filename" : "wmb500.png",
-      "scale" : "3x"
-    }
-  ],
-  "info" : {
-    "version" : 1,
-    "author" : "xcode"
-  }
-}
