+++ trainer/requirements.txt
@@ -1,0 +1,22 @@
+certifi==2019.6.16
+chardet==3.0.4
+coremltools==3.0b3
+decorator==4.4.0
+graphviz==0.8.4
+idna==2.8
+llvmlite==0.29.0
+mxnet==1.1.0.post0
+numba==0.45.1
+numpy==1.16.4
+pandas==0.25.1
+Pillow==6.1.0
+prettytable==0.7.2
+protobuf==3.9.1
+python-dateutil==2.8.0
+pytz==2019.2
+requests==2.22.0
+resampy==0.2.1
+scipy==1.3.1
+six==1.12.0
+turicreate==5.7.1
+urllib3==1.25.3
