+++ wmb/Assets.xcassets/AppIcon.appiconset/Contents.json
@@ -1,0 +1,158 @@
+{
+  "images" : [
+    {
+      "size" : "40x40",
+      "idiom" : "ipad",
+      "filename" : "icon-40.png",
+      "scale" : "1x"
+    },
+    {
+      "size" : "40x40",
+      "filename" : "icon-40@2x.png",
+      "scale" : "2x",
+      "idiom" : "ipad"
+    },
+    {
+      "size" : "60x60",
+      "filename" : "icon-60@2x.png",
+      "idiom" : "iphone",
+      "scale" : "2x"
+    },
+    {
+      "filename" : "icon-72.png",
+      "scale" : "1x",
+      "size" : "72x72",
+      "idiom" : "ipad"
+    },
+    {
+      "filename" : "icon-72@2x.png",
+      "idiom" : "ipad",
+      "size" : "72x72",
+      "scale" : "2x"
+    },
+    {
+      "size" : "76x76",
+      "scale" : "1x",
+      "filename" : "icon-76.png",
+      "idiom" : "ipad"
+    },
+    {
+      "idiom" : "ipad",
+      "scale" : "2x",
+      "filename" : "icon-76@2x.png",
+      "size" : "76x76"
+    },
+    {
+      "filename" : "icon-small-50.png",
+      "scale" : "1x",
+      "size" : "50x50",
+      "idiom" : "ipad"
+    },
+    {
+      "filename" : "icon-small-50@2x.png",
+      "size" : "50x50",
+      "scale" : "2x",
+      "idiom" : "ipad"
+    },
+    {
+      "idiom" : "iphone",
+      "filename" : "icon-small.png",
+      "size" : "29x29",
+      "scale" : "1x"
+    },
+    {
+      "filename" : "icon-small@2x.png",
+      "idiom" : "iphone",
+      "size" : "29x29",
+      "scale" : "2x"
+    },
+    {
+      "filename" : "icon.png",
+      "size" : "57x57",
+      "scale" : "1x",
+      "idiom" : "iphone"
+    },
+    {
+      "size" : "57x57",
+      "filename" : "icon@2x.png",
+      "scale" : "2x",
+      "idiom" : "iphone"
+    },
+    {
+      "idiom" : "iphone",
+      "scale" : "3x",
+      "filename" : "icon-small@3x.png",
+      "size" : "29x29"
+    },
+    {
+      "idiom" : "iphone",
+      "filename" : "icon-40@3x.png",
+      "scale" : "3x",
+      "size" : "40x40"
+    },
+    {
+      "size" : "60x60",
+      "idiom" : "iphone",
+      "scale" : "3x",
+      "filename" : "icon-60@3x.png"
+    },
+    {
+      "size" : "40x40",
+      "scale" : "2x",
+      "idiom" : "iphone",
+      "filename" : "icon-40@2x.png"
+    },
+    {
+      "size" : "29x29",
+      "filename" : "icon-small.png",
+      "idiom" : "ipad",
+      "scale" : "1x"
+    },
+    {
+      "filename" : "icon-small@2x.png",
+      "size" : "29x29",
+      "scale" : "2x",
+      "idiom" : "ipad"
+    },
+    {
+      "filename" : "icon-83.5@2x.png",
+      "idiom" : "ipad",
+      "scale" : "2x",
+      "size" : "83.5x83.5"
+    },
+    {
+      "size" : "20x20",
+      "idiom" : "iphone",
+      "scale" : "2x",
+      "filename" : "notification-icon@2x.png"
+    },
+    {
+      "size" : "20x20",
+      "scale" : "3x",
+      "idiom" : "iphone",
+      "filename" : "notification-icon@3x.png"
+    },
+    {
+      "size" : "20x20",
+      "idiom" : "ipad",
+      "scale" : "1x",
+      "filename" : "notification-icon~ipad.png"
+    },
+    {
+      "scale" : "2x",
+      "idiom" : "ipad",
+      "filename" : "notification-icon~ipad@2x.png",
+      "size" : "20x20"
+    },
+    {
+      "filename" : "ios-marketing.png",
+      "idiom" : "ios-marketing",
+      "scale" : "1x",
+      "size" : "1024x1024"
+    }
+  ],
+  "info" : {
+    "version" : 1,
+    "author" : "xcode"
+  }
+}
